pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "feedparser>=6.0.11",
    "google-genai>=1.29.0",
    "lxml>=5.4.0",
    "pillow>=11.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-slugify>=8.0.4",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
    "urllib3>=2.5.0",
]
